style: split ast into multiple modules
